Refactor Serializer.
git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@3516 86035c5f-d46c-4136-a2d5-fe63f1d4a782

+++ branches/ud/sw/include/component_manager.h
@@ -47,7 +47,7 @@ public:
     static void init_buffer();
 
     static void call(Id id, const Method m, unsigned int n_args,
-        unsigned int n_ret, unsigned int * data);
+        unsigned int n_ret, unsigned long * data);
 
 private:
     static void int_handler(const unsigned int & interrupt);
